Demodulize annotation models
