--- app/models/annotations/conversation.rb
@@ -1,40 +1,0 @@
-# frozen_string_literal: true
-
-module Annotations
-  ##
-  # A conversation (comment thread starter)
-  #
-  class Conversation < Annotation
-    ##
-    # Properties
-    #
-    enum :comment_type => %i[question note]
-
-    property :text
-
-    ##
-    # Associations
-    #
-    has_many :comments,
-             :dependent => :destroy,
-             :inverse_of => :conversation
-
-    ##
-    # Validations
-    #
-    validates :text, :presence => true
-
-    ##
-    # Callbacks
-    #
-    ##
-    # Methods
-    #
-    ##
-    # Overrides
-    #
-    ##
-    # Helpers and callback methods
-    #
-  end
-end
